Split database to different stores
