+++ packages/nexrender-server/package.json
@@ -22,7 +22,8 @@
     "chalk": "^2.4.2",
     "micro": "^9.3.3",
     "micro-cors": "^0.1.1",
-    "microrouter": "^3.1.3"
+    "microrouter": "^3.1.3",
+    "redis": "^3.1.2"
   },
   "publishConfig": {
     "access": "public"
